requirements.txt
streamlit
google-cloud-firestore
json
request
wikipedia
uuid